fix: avoid cargo build from failing due to no main on non-linux plats
